Add SessionService to handle session business logic
