+++ src/main/java/com/example/echo_api/service/session/SessionService.java
@@ -1,0 +1,9 @@
+package com.example.echo_api.service.session;
+
+import com.example.echo_api.persistence.model.User;
+
+public interface SessionService {
+
+    public User getAuthenticatedUser();
+
+}
